backend/requirements.txt
Cython==0.29
Django==1.11.16
django-extensions==2.1.3
django-annoying==0.10.4
djangorestframework==3.9.0
django-two-factor-auth==1.6.2
psycopg2-binary==2.8.3
python-dateutil==2.7.5
pytz==2018.7
pylibmc==1.6.0
python3-saml==1.4.1
path.py==11.5.0
requests==2.20.1
ipython==7.1.1
ipdb==0.11
uWSGI==2.0.17.1
whitenoise==4.1.1
pyOpenSSL==18.0.0
Markdown==3.0.1
mistune==0.8.4
raven==6.9.0
raven-sh==0.4

## The following requirements were added by pip freeze:
asn1crypto==0.24.0
Babel==2.6.0
backcall==0.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.1
decorator==4.3.0
defusedxml==0.5.0
django-formtools==2.1
django-otp==0.5.1
django-phonenumber-field==1.3.0
idna==2.7
importlib-metadata==0.6
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.13.1
lxml==4.2.5
parso==0.3.1
pexpect==4.6.0
phonenumberslite==8.9.16
pickleshare==0.7.5
pkgconfig==1.4.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
qrcode==4.0.4
six==1.11.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
xmlsec==1.3.3
